Move visualization method to the newly created graphviz file
