+++ ruby/binary_search_trees/binary_search_trees.rb
@@ -1,76 +1,3 @@
-# THESE METHODS RESEMBLE TEXTBOOK PSEUDO CODE OR PROFESSOR CALLAHAN'S PYTHON CODE
-
-#      tree_insert
-#      tree_delete -> transplant
-#
-
-# Comment the below line while running the tests. Installation steps for ruby-graphviz
-# and GraphViz will be added soon to a readme file
-require 'ruby-graphviz'
-# Public: Analogous to a struct in C/C++ for building linked lists.
-# This class only contains an initialize method which acts a constructor for
-# setting and accessing the object properties
-#
-# There are multiple ways this can be defined, this is my personal favorite
-# Other ways are listed at the bottom of the file
-# Choose whatever is convenient for YOU.
-#
-# Examples
-#   Tree.new(nil)
-#   # => #<Tree:0x007ffeab2187a0 @root=nil>
-class Tree
-  attr_accessor :root
-
-  def initialize(root)
-    @root = root
-  end
-end
-
-# Public: Analogous to a struct in C/C++ for building linked lists.
-# This class only contains an initialize method which acts a constructor for
-# setting and accessing the object properties
-#
-# NOTE: Two constructors, one with satellite data and another without. The later
-#       one will be used predominantly
-#
-# Examples
-#   Node.new(10, "HELLO", nil, nil, nil)
-#   # => #<Node:0x007ffeab2187a0 @key=10, @satellite_data="HELLO", @p=nil, @left=nil, @right=nil>
-class Node
-  attr_accessor :key, :satellite_data, :p, :left, :right
-  def initialize(key, satellite_data, p, left, right)
-    @key, @satellite_data, @p, @left, @right = key, satellite_data, p, left, right
-  end
-
-  #   Node.new(10, nil, nil, nil)
-  #   # => #<Node:0x007ffeab2187a0 @key=10, @p=nil, @left=nil, @right=nil>
-  def initialize(key, p, left, right)
-    @key, @p, @left, @right = key, p, left, right
-  end
-end
-
-#             TREE structure
-#                  F
-#                /   \
-#              B      G
-#            /  \      \
-#          A     D      I
-#              /  \    /
-#             C   E   H
-#   tree = Tree.new(F)   NOTE: Whole tree is assumed to have been constructed alread
-
-# TEST inorder_tree_walk
-#      preoder_tree_walk
-#      postorder_tree_walk
-#      tree_search
-#      iterative_tree_search
-#      tree_minimum
-#      tree_maximum
-#      tree_successor
-#      tree_predecessor
-#      tree_insert
-#      tree_delete
-#      tree_delete_predecessor
 def unit_test_binary_tree
   node_a = Node.new('A', nil, nil, nil)
   node_b = Node.new('B', nil, nil, nil)
@@ -194,39 +121,3 @@ def unit_test_binary_tree
   p "--------------------  END NODE DELETION (PREDECESSOR) -------------------- "
   preoder_tree_walk(tree.root)
 end
-
-def build_visualization_images
-  node_a = Node.new('A', nil, nil, nil)
-  node_b = Node.new('B', nil, nil, nil)
-  node_c = Node.new('C', nil, nil, nil)
-  node_d = Node.new('D', nil, nil, nil)
-  node_e = Node.new('E', nil, nil, nil)
-  node_f = Node.new('F', nil, nil, nil)
-  node_g = Node.new('G', nil, nil, nil)
-  node_h = Node.new('H', nil, nil, nil)
-  node_i = Node.new('I', nil, nil, nil)
-  node_j = Node.new('J', nil, nil, nil)
-
-  tree_graphical = Tree.new(nil)
-
-  g = GraphViz.new( :G, :type => :digraph )
-
-  p "--------------------  BEGIN GENERATING IMAGES -------------------- "
-  [node_f, node_b, node_g, node_a, node_d, node_c, node_e, node_i, node_h].each_with_index do |x, i|
-    tree_insert_graphical(tree_graphical, x, g) # For image generation
-    g.output( :png => "Order #{i.to_i} - Inserted #{x.key}.png", :canon => nil)
-  end
-  p "                              F                      "
-  p "                            // \\                    "
-  p "                           B     G                   "
-  p "                         // \\    \\                 "
-  p "                         A   D     I                 "
-  p "                           // \\ //                  "
-  p "                           C   E H                   "
-
-  p "--------------------  END GENERATING IMAGES -------------------- "
-end
-
-# Uncomment the lines below to run the minimal unit tests
-unit_test_binary_tree
-build_visualization_images
